move handler code to dedicated folder
