+++ api/handler/user.go
@@ -1,6 +1,8 @@
-package api
+package handler
 
 import (
+	"crypto/rand"
+	"encoding/hex"
 	"net/http"
 
 	"github.com/gin-contrib/sessions"
@@ -9,13 +11,15 @@ import (
 
 func Login(ctx *gin.Context) {
 	session := sessions.Default(ctx)
-	// id & email didapat darimana?
-	id := 122323
+	// token & email didapat darimana?
+
 	email := "mriansyah93@gmail.com"
 
-	session.Set("id", id)
-	session.Set("email", email)
+	// userid akan dipakai untuk mencatat recently view
+	// userId :=
+	session.Set("userId", userId)
 	session.Save()
+
 	ctx.JSON(http.StatusOK, gin.H{
 		"message": "User Sign In successfully",
 	})
@@ -28,4 +32,12 @@ func Logout(ctx *gin.Context) {
 	ctx.JSON(http.StatusOK, gin.H{
 		"message": "success logout",
 	})
+}
+
+func generateSecureToken(length int) string {
+	b := make([]byte, length)
+	if _, err := rand.Read(b); err != nil {
+		return ""
+	}
+	return hex.EncodeToString(b)
 }
