+++ api/router.go
@@ -1,11 +1,18 @@
 package api
 
 import (
+	"sf-duplicate/api/handler"
+	"sf-duplicate/repository"
+
 	"github.com/gin-contrib/sessions"
+	"github.com/gin-contrib/sessions/redis"
 	"github.com/gin-gonic/gin"
 )
 
-func GetRouter(sfapi *SfApi) *gin.Engine {
+func StartRouter(arRepo *repository.ArticleRepo) {
+	arHandler := handler.NewArticleHandler(arRepo)
+	usrHandler := handler.NewUserHandler(arRepo)
+
 	router := gin.Default()
 	router.Use(sessions.Sessions("sfsession", getRedisStore()))
 
@@ -22,9 +29,9 @@ func GetRouter(sfapi *SfApi) *gin.Engine {
 	user.Use(sessionAuth())
 
 	{
-		user.GET("/home", sfapi.listArticles)
+		user.GET("/home", arHandler.listArticles)
 	}
-	return router
+	router.Run(":8080")
 }
 
 func rootHandler(ctx *gin.Context) {
@@ -33,10 +40,19 @@ func rootHandler(ctx *gin.Context) {
 	})
 }
 
+func getRedisStore() redis.Store {
+	pwd := []byte("secret")
+	size := 10
+	redisStore, _ := redis.NewStore(size, "tcp", "localhost:6379", "", pwd)
+	return redisStore
+}
+
 // try redis session
 func incrementHandler(ctx *gin.Context) {
 	var counter int
 	session := sessions.Default(ctx)
+
+	token := session.Get("token")
 	val := session.Get("counter")
 
 	if val == nil {
@@ -46,9 +62,11 @@ func incrementHandler(ctx *gin.Context) {
 		counter++
 	}
 	session.Set("counter", counter)
+	session.Set("token", generateSecureToken(10))
 	session.Save()
 
 	ctx.JSON(200, gin.H{
 		"counter": counter,
+		"token":   token,
 	})
 }
